@@ -10,7 +10,7 @@
     }
   ],
   "require": {
-    "php": ">=8.0.2",
+    "php": ">=8.1",
     "ext-fileinfo": "*",
     "ext-zip": "*",
     "ext-openssl": "*"
@@ -24,7 +24,7 @@
   },
   "config": {
     "platform": {
-      "php": "8.0.2"
+      "php": "8.1"
     },
     "sort-packages": true
   },
