@@ -186,36 +186,25 @@ class Memory extends Storage
     /**
      * @inheritDoc
      */
-    public function getFileHash(int $mode = Hash::CONTENT, string $algo = 'sha256', bool $raw = false): string
+    public function getFileHash(Hash $mode = Hash::CONTENT, string $algo = 'sha256', bool $raw = false): string
     {
-        switch ($mode) {
-            case Hash::CONTENT:
-                return hash($algo, $this->content ?? '', $raw);
-            case Hash::LAST_MODIFIED:
-                return hash($algo, (string)$this->lastModified, $raw);
-            case Hash::FILENAME:
-            case Hash::FILEPATH:
-                throw new RuntimeException('unable to calculate filepath/name hash for in-memory-files', 500);
-            default:
-                throw new RuntimeException('unknown hashing-mode', 500);
-        }
+        return match ($mode) {
+            Hash::CONTENT => hash($algo, $this->content ?? '', $raw),
+            Hash::LAST_MODIFIED => hash($algo, (string)$this->lastModified, $raw),
+            default => throw new RuntimeException('unable to calculate filepath/name hash for in-memory-files', 500),
+        };
     }
 
     /**
      * @inheritDoc
      */
-    public function getTime(int $type = Time::LAST_MODIFIED): ?int
+    public function getTime(Time $type = Time::LAST_MODIFIED): ?int
     {
-        switch ($type) {
-            case Time::LAST_MODIFIED:
-                return $this->lastModified;
-            case Time::LAST_ACCESSED:
-                return $this->lastAccessed;
-            case Time::CREATED:
-                return $this->created;
-        }
-
-        return null;
+        return match ($type) {
+            Time::LAST_MODIFIED => $this->lastModified,
+            Time::LAST_ACCESSED => $this->lastAccessed,
+            Time::CREATED => $this->created,
+        };
     }
 
     /**
@@ -319,7 +308,7 @@ class Memory extends Storage
                 if (!$destination->writeFile($this->readFile())) {
                     return false;
                 }
-                $destination->path()->reload();
+                $destination->getPath()->reload();
                 return true;
 
             case $destination instanceof Flysystem:
