@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace ricwein\FileSystem\Enum;
 
-class Time
+enum Time: int
 {
-    public const LAST_MODIFIED = 0b0001;
-    public const LAST_ACCESSED = 0b0010;
-    public const CREATED = 0b0100;
+    case LAST_MODIFIED = 0b0001;
+    case LAST_ACCESSED = 0b0010;
+    case CREATED = 0b0100;
 }
