@@ -1,9 +1,4 @@
 <?php
-
-/**
- * @author Richard Weinhold
- */
-
 declare(strict_types=1);
 
 namespace ricwein\FileSystem;
@@ -45,7 +40,7 @@ class File extends FileSystem
     protected function checkFileReadPermissions(): void
     {
         if (!$this->isFile() || !$this->isReadable()) {
-            throw new FileNotFoundException(sprintf('unable to open file: "%s"', $this->storage instanceof Storage\Disk ? $this->storage->path()->raw : get_class($this->storage)), 404);
+            throw new FileNotFoundException(sprintf('unable to open file: "%s"', $this->storage instanceof Storage\Disk ? $this->storage->getPath()->getRealOrRawPath() : get_class($this->storage)), 404);
         }
 
         parent::checkFileReadPermissions(); // TODO: Change the autogenerated stub
@@ -59,11 +54,11 @@ class File extends FileSystem
     protected function checkFileWritePermissions(): void
     {
         if (!$this->storage->doesSatisfyConstraints()) {
-            throw new AccessDeniedException(sprintf('unable to write file: "%s"', $this->storage instanceof Storage\Disk ? $this->storage->path()->raw : get_class($this->storage)), 403, $this->storage->getConstraintViolations());
+            throw new AccessDeniedException(sprintf('unable to write file: "%s"', $this->storage instanceof Storage\Disk ? $this->storage->getPath()->getRawPath() : get_class($this->storage)), 403, $this->storage->getConstraintViolations());
         }
 
         if ($this->isFile() && !$this->isWriteable()) {
-            throw new AccessDeniedException(sprintf('unable to write file: "%s"', $this->storage instanceof Storage\Disk ? $this->storage->path()->raw : get_class($this->storage)), 403);
+            throw new AccessDeniedException(sprintf('unable to write file: "%s"', $this->storage instanceof Storage\Disk ? $this->storage->getPath()->getRawPath() : get_class($this->storage)), 403);
         }
     }
 
@@ -166,7 +161,7 @@ class File extends FileSystem
         // ensure the destination-path points to a filename
         if ($destination instanceof Storage\Disk && $destination->isDir()) {
             $destination = clone $destination;
-            $destination->cd([$this->storage instanceof Storage\Disk ? $this->path()->filename : (uniqid('', true) . '.file')]);
+            $destination->cd([$this->storage instanceof Storage\Disk ? $this->getPath()->getFilename() : (uniqid('', true) . '.file')]);
         }
 
         // actual copy file to file: use native functions if possible
@@ -225,7 +220,7 @@ class File extends FileSystem
         // ensure the destination-path points to a filename
         if ($destination instanceof Storage\Disk && $destination->isDir()) {
             $destination = clone $destination;
-            $destination->cd([$this->storage instanceof Storage\Disk ? $this->path()->filename : (uniqid('', true) . '.file')]);
+            $destination->cd([$this->storage instanceof Storage\Disk ? $this->getPath()->getFilename() : (uniqid('', true) . '.file')]);
         }
 
         // actual move file to file: use native functions if possible
@@ -269,7 +264,7 @@ class File extends FileSystem
      * @throws RuntimeException
      * @throws UnexpectedValueException
      */
-    public function getHash(int $mode = Hash::CONTENT, string $algo = 'sha256', bool $raw = false): string
+    public function getHash(Hash $mode = Hash::CONTENT, string $algo = 'sha256', bool $raw = false): string
     {
         $this->checkFileReadPermissions();
 
@@ -318,7 +313,7 @@ class File extends FileSystem
 
         // validate constraints
         if (!$this->isFile()) {
-            throw new FileNotFoundException(sprintf('unable to open file: "%s"', $this->storage instanceof Storage\Disk ? $this->storage->path()->raw : get_class($this->storage)), 404);
+            throw new FileNotFoundException(sprintf('unable to open file: "%s"', $this->storage instanceof Storage\Disk ? $this->storage->getPath()->getRawPath() : get_class($this->storage)), 404);
         }
 
         if (!$this->storage->removeFile()) {
@@ -342,10 +337,9 @@ class File extends FileSystem
 
     /**
      * @throws ConstraintsException
+     * @throws FilesystemException
      * @throws RuntimeException
      * @throws UnexpectedValueException
-     * @throws Exceptions\UnsupportedException
-     * @throws FilesystemException
      */
     public function dir(?int $constraints = null, string $as = Directory::class, mixed ...$arguments): Directory
     {
@@ -354,7 +348,7 @@ class File extends FileSystem
         }
 
         if ($this->storage instanceof Storage\Flysystem) {
-            $filePath = $this->storage->path();
+            $filePath = $this->storage->getPath()->getRawPath();
 
             if (false !== $lastDelimiterPos = strrpos($filePath, '/')) {
                 $dirPath = substr($filePath, 0, $lastDelimiterPos);
@@ -370,18 +364,18 @@ class File extends FileSystem
         }
 
 
-        $path = $this->path();
-        $dirPath = $path->directory;
+        $path = $this->getPath();
+        $dirPath = $path->getDirectory();
         if (is_dir($dirPath)) {
             $dirPath = realpath($dirPath);
         }
 
-        $safepath = $path->safepath;
+        $safepath = $path->getSafePath();
         if (is_dir($safepath)) {
             $safepath = realpath($safepath);
         }
 
-        if ($safepath !== null && is_dir($safepath)) {
+        if ($safepath !== false && is_dir($safepath)) {
             $storage = new Storage\Disk($safepath, str_replace($safepath, '', $dirPath));
         } else {
             $storage = new Storage\Disk($dirPath);
