@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace ricwein\FileSystem\Tests\Directory;
 
+use League\Flysystem\FilesystemException;
 use PHPUnit\Framework\TestCase;
 use ricwein\FileSystem\Exceptions\AccessDeniedException;
 use ricwein\FileSystem\Exceptions\ConstraintsException;
@@ -11,6 +12,7 @@ use ricwein\FileSystem\Exceptions\Exception;
 use ricwein\FileSystem\Exceptions\FileNotFoundException;
 use ricwein\FileSystem\Exceptions\RuntimeException;
 use ricwein\FileSystem\Exceptions\UnexpectedValueException;
+use ricwein\FileSystem\Exceptions\UnsupportedException;
 use ricwein\FileSystem\File;
 use ricwein\FileSystem\Storage;
 use ricwein\FileSystem\Directory;
@@ -23,37 +25,40 @@ class NavigationTest extends TestCase
      * @throws AccessDeniedException
      * @throws RuntimeException
      * @throws UnexpectedValueException
+     * @throws UnsupportedException
      */
     public function testDirChange(): void
     {
         $dir = new Directory(new Storage\Disk(__DIR__ . '/../'));
-        self::assertSame($dir->path()->real, realpath(__DIR__ . '/../'));
+        self::assertSame($dir->getPath()->getRealPath(), realpath(__DIR__ . '/../'));
 
         $dir->cd('_examples');
-        self::assertSame($dir->path()->real, realpath(__DIR__ . '/../_examples/'));
+        self::assertSame($dir->getPath()->getRealPath(), realpath(__DIR__ . '/../_examples/'));
     }
 
     /**
      * @throws AccessDeniedException
      * @throws RuntimeException
      * @throws UnexpectedValueException
+     * @throws UnsupportedException
      */
     public function testDirUp(): void
     {
         $dir = new Directory(new Storage\Disk(__DIR__));
-        self::assertSame($dir->path()->real, realpath(__DIR__));
+        self::assertSame($dir->getPath()->getRealPath(), realpath(__DIR__));
 
         $dir->up(2);
-        self::assertSame($dir->path()->real, realpath(__DIR__ . '/../../'));
+        self::assertSame($dir->getPath()->getRealPath(), realpath(__DIR__ . '/../../'));
     }
 
     /**
      * @throws AccessDeniedException
      * @throws ConstraintsException
+     * @throws FileNotFoundException
      * @throws RuntimeException
      * @throws UnexpectedValueException
-     * @throws Exception
-     * @throws FileNotFoundException
+     * @throws UnsupportedException
+     * @throws FilesystemException
      */
     public function testFileRead(): void
     {
@@ -79,14 +84,15 @@ class NavigationTest extends TestCase
      * @throws FileNotFoundException
      * @throws RuntimeException
      * @throws UnexpectedValueException
+     * @throws FilesystemException
      */
     public function testDirUpError(): void
     {
         $dir = new Directory(new Storage\Disk(__DIR__));
-        self::assertSame($dir->path()->real, realpath(__DIR__));
+        self::assertSame($dir->getPath()->getRealPath(), realpath(__DIR__));
 
         $this->expectException(ConstraintsException::class);
-        $this->expectExceptionMessageMatches('/.*constraint failed: the given real-path.*is not within the safepath.*/');
+        $this->expectExceptionMessageMatches('/.*constraint failed: the given path (.*) is not within the safepath (.*)$/');
 
         $dir->up(2);
         $dir->file('LICENSE')->read();
