@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 namespace ricwein\FileSystem\Tests\Directory;
 
+use League\Flysystem\FilesystemException;
 use PHPUnit\Framework\TestCase;
 use ricwein\FileSystem\Directory;
 use ricwein\FileSystem\Exceptions\AccessDeniedException;
 use ricwein\FileSystem\Exceptions\ConstraintsException;
 use ricwein\FileSystem\Exceptions\RuntimeException;
 use ricwein\FileSystem\Exceptions\UnexpectedValueException;
+use ricwein\FileSystem\Exceptions\UnsupportedException;
 use ricwein\FileSystem\Helper\Constraint;
 use ricwein\FileSystem\Storage;
 
@@ -17,22 +19,23 @@ class WriteTest extends TestCase
 {
     /**
      * @throws AccessDeniedException
-     * @throws RuntimeException
-     * @throws UnexpectedValueException
+     * @throws UnsupportedException
      */
     public function testCreateDir(): void
     {
         $dir = new Directory(new Storage\Disk\Temp());
-        self::assertFileExists($dir->path()->raw);
-        self::assertFileExists($dir->path()->real);
-        self::assertDirectoryExists($dir->path()->real);
+        self::assertFileExists($dir->getPath()->getRawPath());
+        self::assertFileExists($dir->getPath()->getRealPath());
+        self::assertDirectoryExists($dir->getPath()->getRealPath());
     }
 
     /**
      * @throws AccessDeniedException
+     * @throws ConstraintsException
      * @throws RuntimeException
      * @throws UnexpectedValueException
-     * @throws ConstraintsException
+     * @throws UnsupportedException
+     * @throws FilesystemException
      */
     public function testCreateRecursiveDir(): void
     {
@@ -44,29 +47,28 @@ class WriteTest extends TestCase
         $dir3->mkdir();
 
         foreach ([$dir1, $dir2, $dir3] as $dir) {
-            self::assertFileExists($dir->path()->raw);
-            self::assertFileExists($dir->path()->real);
-            self::assertDirectoryExists($dir->path()->real);
+            self::assertFileExists($dir->getPath()->getRawPath());
+            self::assertFileExists($dir->getPath()->getRealPath());
+            self::assertDirectoryExists($dir->getPath()->getRealPath());
         }
 
-        self::assertSame('/dir2', str_replace($dir1->path()->real, '', $dir2->path()->real));
-        self::assertSame('/dir3/dir4/dir5', str_replace($dir1->path()->real, '', $dir3->path()->real));
+        self::assertSame('/dir2', str_replace($dir1->getPath()->getRealPath(), '', $dir2->getPath()->getRealPath()));
+        self::assertSame('/dir3/dir4/dir5', str_replace($dir1->getPath()->getRealPath(), '', $dir3->getPath()->getRealPath()));
     }
 
     /**
      * @throws AccessDeniedException
-     * @throws RuntimeException
-     * @throws UnexpectedValueException
+     * @throws UnsupportedException
      */
     public function testTempDirRemoval(): void
     {
         $tmpDir = new Directory(new Storage\Disk\Temp());
 
-        self::assertFileExists($tmpDir->path()->raw);
-        self::assertFileExists($tmpDir->path()->real);
-        self::assertDirectoryExists($tmpDir->path()->real);
+        self::assertFileExists($tmpDir->getPath()->getRawPath());
+        self::assertFileExists($tmpDir->getPath()->getRealPath());
+        self::assertDirectoryExists($tmpDir->getPath()->getRealPath());
 
-        $path = $tmpDir->path()->raw;
+        $path = $tmpDir->getPath()->getRawPath();
         $tmpDir = null;
 
         self::assertFileDoesNotExist($path);
