@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace ricwein\FileSystem\Tests\File;
 
+use League\Flysystem\FilesystemException;
 use PHPUnit\Framework\TestCase;
 use ricwein\FileSystem\Exceptions\AccessDeniedException;
 use ricwein\FileSystem\Exceptions\ConstraintsException;
@@ -11,6 +12,7 @@ use ricwein\FileSystem\Exceptions\Exception;
 use ricwein\FileSystem\Exceptions\FileNotFoundException;
 use ricwein\FileSystem\Exceptions\RuntimeException;
 use ricwein\FileSystem\Exceptions\UnexpectedValueException;
+use ricwein\FileSystem\Exceptions\UnsupportedException;
 use ricwein\FileSystem\File;
 use ricwein\FileSystem\Storage;
 use ricwein\FileSystem\Directory;
@@ -25,6 +27,7 @@ class CopyTest extends TestCase
      * @throws FileNotFoundException
      * @throws RuntimeException
      * @throws UnexpectedValueException
+     * @throws FilesystemException
      */
     public function testCopyFromDiskToDisk(): void
     {
@@ -47,6 +50,7 @@ class CopyTest extends TestCase
      * @throws ConstraintsException
      * @throws Exception
      * @throws FileNotFoundException
+     * @throws FilesystemException
      * @throws RuntimeException
      * @throws UnexpectedValueException
      */
@@ -69,6 +73,7 @@ class CopyTest extends TestCase
      * @throws ConstraintsException
      * @throws Exception
      * @throws FileNotFoundException
+     * @throws FilesystemException
      * @throws RuntimeException
      * @throws UnexpectedValueException
      */
@@ -88,6 +93,7 @@ class CopyTest extends TestCase
      * @throws ConstraintsException
      * @throws Exception
      * @throws FileNotFoundException
+     * @throws FilesystemException
      * @throws RuntimeException
      * @throws UnexpectedValueException
      */
@@ -107,8 +113,10 @@ class CopyTest extends TestCase
      * @throws ConstraintsException
      * @throws Exception
      * @throws FileNotFoundException
+     * @throws FilesystemException
      * @throws RuntimeException
      * @throws UnexpectedValueException
+     * @throws UnsupportedException
      */
     public function testCopyToDir(): void
     {
@@ -120,8 +128,8 @@ class CopyTest extends TestCase
 
         $retFile = $source->copyTo($destination->storage());
 
-        self::assertSame($source->path()->filename, $retFile->path()->filename);
-        self::assertSame($destination->path()->raw, $retFile->path()->directory);
+        self::assertSame($source->getPath()->getFilename(), $retFile->getPath()->getFilename());
+        self::assertSame($destination->getPath()->getRealPath(), $retFile->getPath()->getDirectory());
         self::assertStringEqualsFile(__DIR__ . '/../_examples/test.txt', $retFile->read());
     }
 
@@ -130,8 +138,10 @@ class CopyTest extends TestCase
      * @throws ConstraintsException
      * @throws Exception
      * @throws FileNotFoundException
+     * @throws FilesystemException
      * @throws RuntimeException
      * @throws UnexpectedValueException
+     * @throws UnsupportedException
      */
     public function testCopyMemoryToDir(): void
     {
@@ -143,8 +153,8 @@ class CopyTest extends TestCase
 
         $retFile = $source->copyTo($destination->storage());
 
-        self::assertNotFalse(strpos($retFile->path()->filename, '.file'));
-        self::assertSame($destination->path()->raw, $retFile->path()->directory);
+        self::assertNotFalse(strpos($retFile->getPath()->getFilename(), '.file'));
+        self::assertSame($destination->getPath()->getRealPath(), $retFile->getPath()->getDirectory());
         self::assertStringEqualsFile(__DIR__ . '/../_examples/test.txt', $retFile->read());
     }
 }
