@@ -7,41 +7,44 @@ namespace ricwein\FileSystem\Tests\Helper;
 use PHPUnit\Framework\TestCase;
 use ricwein\FileSystem\Exceptions\RuntimeException;
 use ricwein\FileSystem\Exceptions\UnexpectedValueException;
-use ricwein\FileSystem\Exceptions\UnsupportedException;
 use ricwein\FileSystem\Helper\Constraint;
-use ricwein\FileSystem\Helper\Path;
+use ricwein\FileSystem\Path;
 
 /**
  * @author Richard Weinhold
  */
 class ConstraintTest extends TestCase
 {
-    /**
-     * @throws RuntimeException
-     * @throws UnexpectedValueException
-     * @throws UnsupportedException
-     */
-    public function testPathConstraints(): void
+    public function testPathConstraintsSimple(): void
     {
         // safe-path
-        $path = new Path([realpath(__DIR__ . '/../_examples'), 'test.txt']);
+        $path = new Path(__FILE__);
+        self::assertTrue((new Constraint(Constraint::IN_SAFEPATH | Constraint::DISALLOW_LINK))->isValidPath($path));
+    }
+
+    /**
+     * @throws UnexpectedValueException
+     * @throws RuntimeException
+     */
+    public function testPathConstraintsSafe(): void
+    {
+        // safe-path
+        $path = new Path(realpath(__DIR__ . '/../_examples'), 'test.txt');
 
         self::assertTrue((new Constraint(Constraint::DISALLOW_LINK))->isValidPath($path));
         self::assertTrue((new Constraint(Constraint::IN_SAFEPATH))->isValidPath($path));
         self::assertTrue((new Constraint(Constraint::IN_SAFEPATH | Constraint::DISALLOW_LINK))->isValidPath($path));
-        self::assertTrue((new Constraint(Constraint::STRICT & ~Constraint::IN_OPENBASEDIR))->isValidPath($path));
+        self::assertTrue((new Constraint(Constraint::STRICT & ~Constraint::IN_OPEN_BASEDIR))->isValidPath($path));
+    }
 
+    public function testPathConstraintsUnsafe(): void
+    {
         // unsafe-path
-        $path = new Path([__DIR__, '/../', '_examples', 'test.txt']);
+        $path = new Path(__DIR__, '/../', '_examples', 'test.txt');
 
         self::assertTrue((new Constraint(Constraint::DISALLOW_LINK))->isValidPath($path));
         self::assertFalse((new Constraint(Constraint::IN_SAFEPATH))->isValidPath($path));
         self::assertFalse((new Constraint(Constraint::IN_SAFEPATH | Constraint::DISALLOW_LINK))->isValidPath($path));
-        self::assertFalse((new Constraint(Constraint::STRICT & ~Constraint::IN_OPENBASEDIR))->isValidPath($path));
-
-        // safe-path again
-        $path = new Path([__FILE__]);
-
-        self::assertTrue((new Constraint(Constraint::IN_SAFEPATH | Constraint::DISALLOW_LINK))->isValidPath($path));
+        self::assertFalse((new Constraint(Constraint::STRICT & ~Constraint::IN_OPEN_BASEDIR))->isValidPath($path));
     }
 }
