@@ -3,12 +3,16 @@ declare(strict_types=1);
 
 namespace ricwein\FileSystem\Tests\Storage;
 
+use League\Flysystem\FilesystemException;
 use PHPUnit\Framework\TestCase;
 use ricwein\FileSystem\Directory;
 use ricwein\FileSystem\Exceptions\AccessDeniedException;
 use ricwein\FileSystem\Exceptions\ConstraintsException;
 use ricwein\FileSystem\Exceptions\Exception;
 use ricwein\FileSystem\Exceptions\FileNotFoundException;
+use ricwein\FileSystem\Exceptions\RuntimeException;
+use ricwein\FileSystem\Exceptions\UnexpectedValueException;
+use ricwein\FileSystem\Exceptions\UnsupportedException;
 use ricwein\FileSystem\File;
 use ricwein\FileSystem\Storage;
 
@@ -65,8 +69,12 @@ class DiskErrorTest extends TestCase
 
     /**
      * @throws AccessDeniedException
-     * @throws Exception
-     * @throws \Exception
+     * @throws ConstraintsException
+     * @throws FileNotFoundException
+     * @throws FilesystemException
+     * @throws RuntimeException
+     * @throws UnexpectedValueException
+     * @throws UnsupportedException
      */
     public function testImplicitDirectoryWriteError(): void
     {
@@ -78,7 +86,7 @@ class DiskErrorTest extends TestCase
     /**
      * @throws AccessDeniedException
      * @throws Exception
-     * @throws \Exception
+     * @throws FilesystemException
      */
     public function testImplicitDirectoryWriteConstraintError(): void
     {
@@ -86,5 +94,4 @@ class DiskErrorTest extends TestCase
         $this->expectException(ConstraintsException::class);
         $dir->remove();
     }
-
 }
